requirements.txt
streamlit
pandas
numpy
openai
scikit-learn
Pillow
requests
python-dotenv
